:art: Total refactoring of pandora_end_effector_controller package
The server can now support any number of clients-modules. Dictionaries
have been added aiding in the accomplishment of this goal.

A factory has been made to automatically produce needed clients.

Classes implementing the clients have also been created, so that
the different proccesses passing the action to the submodules are
implemented